[FLINK-5402] Fails AkkaRpcServiceTest#testTerminationFuture
